Replace Cri with Thor for CLI app
Using Thor to create CLI application is a bit more user-friendly and
extensible than using Cri, but the main reason for the switch is
Thor's parsing of arguments.

Cri is not offering any aids here so we are required to validate the
script arguments on our own. Thor on the other hand will abort on too
few or too many arguments.