Convert body of data store item into hash
Since body of the resource can be modified, it makes sense to keep the
information stored in easily accessible way. This also reduces the
load times, since there is no JSON parsing involved in this step.

The only downside of this change is the need fo body serialization
when returning resource, but this is much simpler and less error prone
process compared to parsing.