Externalize datastore and login_path
Since server is always accessible from the mounted servlet, it makes
little sense to pass data that is common to all requests that are done
to the server down to the servlet on each invocation.

In order to keep the code concise, we simply delegate calls to
datastore and login path to the server.