Implement proper login functionality
Up until now, we did not check credentials at all and always returned
the data. This commit introduces changes that actually check the
username and password that are either sent in basic authorization
header or used to create new session,

In order to fully support session authentication, we added session
token generator and checker, which should make this mock server fully
standard compliant in respect to security.