Add support for firmware update process mocking
The mock is semi-static: we require recording to have update and task
services enabled, but we do not require any static items to be present
in the task collection.

Steps that mocked firmware update process goes through are: New ->
Starting -> Running -> Completed. The process stage is updated on
every GET request to the monitor endpoint. Fetching task instance will
not update the state.

We fully support monitor and task combo that Redfish standard
prescribes. When new update operation is started, we create new task
and associated monitor. We also populate the location headerwith
proper value.