Allow adding new attributes to resources
Up until now, mock server did not allow adding new keys to existing
resources. The main reason for this way the fact that vast majority of
resources have its internal structure defined by the standard.

Unfortunately, this approach made it impossible to implement  more
free-form updates of the resources that are needed when we need to
update things like BIOS settings.

To rectify the situation, we dropped the safeguard. This will make
mock server less strict in some situations, but this is the price we
have to pay for free form updates.